+++ src/sys/kernel/x86/descriptor_tables.cpp
@@ -1,3 +1,4 @@
+#include <kernel/log.h>
 #include <kernel/x86/descriptor_table.h>
 #include <string.h>
 extern "C" void setupGDT(uint32_t);
@@ -7,7 +8,7 @@ kernel::x86::gdt kernel::x86::g_gdt;
 void kernel::x86::gdt::gdt_set_gate(int32_t num, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran) {
     gdt_entries[num].base_low    = (base & 0xFFFF);
     gdt_entries[num].base_middle = (base >> 16) & 0xFF;
-    gdt_entries[num].base_high   = (base >> 24) & 0xFF;
+    gdt_entries[num].base_high   = (uint8_t)((base >> 24) & 0xFF);
 
     gdt_entries[num].limit_low   = (limit & 0xFFFF);
     gdt_entries[num].granularity = (limit >> 16) & 0x0F;
