+++ src/sys/kernel/x86/drivers/pit.cpp
@@ -17,10 +17,10 @@ void pit_timer::tick() {
 uint64_t pit_timer::resolution_ns() { return 1000000000 / resolution_hz; }
 
 void pit_timer::set_phase(unsigned int hz) {
-    int divisor = 1193180 / hz;  // Calculate the divisor
-    outb(0x43, 0x36);            // Command 0x36
-    outb(0x40, divisor & 0xFF);  // Send Divisor Lo
-    outb(0x40, divisor >> 8);    // Send Divisor Hi
+    int divisor = 1193180 / hz;             // Calculate the divisor
+    outb(0x43, 0x36);                       // Command 0x36
+    outb(0x40, (uint8_t)(divisor & 0xFF));  // Send Divisor Lo
+    outb(0x40, (uint8_t)(divisor >> 8));    // Send Divisor Hi
 
     resolution_hz = hz;
 }
